requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
chromadb
google-generativeai
pypdf
langchain-text-splitters
python-multipart
requests
sqlalchemy
psycopg2-binary
pgvector